requirements.txt
# 保客通 (BaokeTong) - AI+ 保险获客系统

# Hermes Agent 核心依赖
# nous-hermes-agent>=0.1.0  # 临时注释：该包不在公共 PyPI 上，需要从私有源安装

# AI 模型 — LiteLLM 统一接口（支持 100+ 模型）
litellm>=1.60.0

# Web 框架
fastapi>=0.109.0
uvicorn>=0.27.0

# 数据库
# PostgreSQL 驱动
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
# SQLAlchemy ORM
sqlalchemy>=2.0.25
# Redis 缓存
redis>=5.0.0
# 向量数据库 Qdrant
qdrant-client>=1.7.0

# 数据加密
cryptography>=42.0.0
python-jose[cryptography]>=3.3.0

# 数据处理
pydantic>=2.5.0
pydantic-settings>=2.0.0
pyyaml>=6.0.1
httpx>=0.26.0
aiohttp>=3.9.0

# 向量嵌入
sentence-transformers>=2.3.0

# 监控与日志
langfuse>=2.20.0

# 测试
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
